Expense work and bug fixing

+++ Point Of Sale/InventoryManagementSystem/ManageInventoryForm.cs
@@ -14,8 +14,12 @@ namespace InventoryManagementSystem
     public partial class ManageInventoryForm : Form
     {
         List<POSItemInfo> mItems = null;
+
         public ManageInventoryForm()
         {
+            Cursor currentCursor = Cursor.Current;
+            Cursor.Current = Cursors.WaitCursor;
+            
             InitializeComponent();
 
             if (!LoginForm.mLoggedInUser.IsAdmin)
@@ -28,9 +32,26 @@ namespace InventoryManagementSystem
                 }
             }
 
-            this.mItems = POSDbUtility.GetAllPOSItems();
-            
+            try
+            {
+                this.mItems = POSDbUtility.GetAllPOSItems();
+            }
+            catch (Exception e)
+            {
+                string errorMsg = POSComonUtility.GetInnerExceptionMessage(e);
+                Cursor.Current = currentCursor;
+                MessageBox.Show(this, "Some error occured in fetching users.\n\n" + errorMsg);
+
+            }
+
+            if (this.mItems == null)
+            {
+                this.mItems = new List<POSItemInfo>();
+            }
+
             this.bsPOSItemInfo.DataSource = this.mItems;
+
+            Cursor.Current = currentCursor;
         }
 
         private void btnShowAll_Click(object sender, EventArgs e)
@@ -52,15 +73,28 @@ namespace InventoryManagementSystem
 
             string itemName = this.tbxItemName.Text;
 
-            List<POSItemInfo> items = this.mItems;
+            List<POSItemInfo> items = new List<POSItemInfo>();
 
-            if (!string.IsNullOrEmpty(itemName))
+            if (string.IsNullOrEmpty(itemName))
+            {
+                items = this.mItems;
+            }
+            else
             {
                 POSItemInfo item = this.mItems.Find(posItem => posItem.Barcode == itemName);
 
                 if (item == null)
                 {
-                    items = this.mItems.FindAll(posItem => posItem.Name.Contains(itemName));
+                    item = this.mItems.Find(posItem => posItem.Id.ToString() == itemName);
+
+                    if (item == null)
+                    {
+                        items = this.mItems.FindAll(posItem => posItem.Name.ToLower().Contains(itemName.ToLower()));
+                    }
+                    else
+                    {
+                        items.Add(item);
+                    }
                 }
                 else
                 {
@@ -83,7 +117,21 @@ namespace InventoryManagementSystem
 
             this.tbxItemName.Text = string.Empty;
 
-            this.mItems = POSDbUtility.GetAllPOSItems();
+            try
+            {
+                this.mItems = POSDbUtility.GetAllPOSItems();
+            }
+            catch (Exception e)
+            {
+                string errorMsg = POSComonUtility.GetInnerExceptionMessage(e);
+                Cursor.Current = current;
+                MessageBox.Show(this, "Some error occured in getting items.\n\n" + errorMsg);
+            }
+
+            if (this.mItems == null)
+            {
+                this.mItems = new List<POSItemInfo>();
+            }
 
             this.bsPOSItemInfo.DataSource = null;
             this.bsPOSItemInfo.DataSource = this.mItems;
@@ -104,18 +152,51 @@ namespace InventoryManagementSystem
                 Cursor current = Cursor.Current;
                 Cursor.Current = Cursors.WaitCursor;
 
-                this.mItems = POSDbUtility.GetAllPOSItems();
-                List<POSItemInfo> items = new List<POSItemInfo>();
-                POSItemInfo item = this.mItems.Find(posItem => posItem.Barcode == itemName);
-
-                if (item == null)
+                try
                 {
-                    items = this.mItems.FindAll(posItem => posItem.Name.ToLower().Contains(itemName.ToLower()));
+                    this.mItems = POSDbUtility.GetAllPOSItems();
+                }
+                catch (Exception e)
+                {
+                    string errorMsg = POSComonUtility.GetInnerExceptionMessage(e);
+                    Cursor.Current = current;
+                    MessageBox.Show(this, "Some error occured in getting items.\n\n" + errorMsg);
+                }
+
+                if (this.mItems == null)
+                {
+                    this.mItems = new List<POSItemInfo>();
+                }
+
+                List<POSItemInfo> items = new List<POSItemInfo>();
+
+                if (string.IsNullOrEmpty(itemName))
+                {
+                    items = this.mItems;
                 }
                 else
                 {
-                    items.Add(item);
+                    POSItemInfo item = this.mItems.Find(posItem => posItem.Barcode == itemName);
+
+                    if (item == null)
+                    {
+                        item = this.mItems.Find(posItem => posItem.Id.ToString() == itemName);
+
+                        if (item == null)
+                        {
+                            items = this.mItems.FindAll(posItem => posItem.Name.ToLower().Contains(itemName.ToLower()));
+                        }
+                        else
+                        {
+                            items.Add(item);
+                        }
+                    }
+                    else
+                    {
+                        items.Add(item);
+                    }
                 }
+                
 
                 this.bsPOSItemInfo.DataSource = null;
                 this.bsPOSItemInfo.DataSource = items;
@@ -192,13 +273,25 @@ namespace InventoryManagementSystem
                 }
                 else
                 {
-                    DialogResult dialogResult = MessageBox.Show(this, "Are you sure you want to delete user?", "Confirmation Dialog", MessageBoxButtons.YesNo);
+                    bool billCreated = itemToDelete.BillItems != null && itemToDelete.BillItems.Count > 0;
+
+                    if (billCreated)
+                    {
+                        MessageBox.Show(this, "This item is already sold, so this item can not be deleted.");
+                        return;
+                    }
+
+                    DialogResult dialogResult = MessageBox.Show(this, "Are you sure you want to delete selected item?", "Confirmation Dialog", MessageBoxButtons.YesNo);
 
                     string errorMsg = string.Empty;
                     if (dialogResult == DialogResult.Yes)
-                    {                        
+                    {
+                        Cursor currentCursor = Cursor.Current;
+                        Cursor.Current = Cursors.WaitCursor;
+                        
                         POSStatusCodes status = POSDbUtility.DeletePOSItem(itemToDelete, ref errorMsg, true);
 
+                        Cursor.Current = currentCursor;
                         if (status == POSStatusCodes.Success)
                         {
                             string itemName = this.tbxItemName.Text;
@@ -215,7 +308,9 @@ namespace InventoryManagementSystem
                         else
                         {
                             MessageBox.Show(this, "Some error occured in deleting Item.\n\n" + errorMsg);
-                        }                        
+                        }
+
+                                  
                     }
                 }
             }
